@@ -51,11 +55,11 @@
     "cors": "^2.8.5",
     "cz-conventional-changelog": "^3.3.0",
     "es6-promise": "^4.2.8",
-    "eslint": "^7.32.0",
+    "eslint": "^8.57.1",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "4.2.1",
+    "eslint-plugin-prettier": "^5.5.4",
     "express": "^4.17.1",
-    "filenamify-url": "1.0.0",
+    "filenamify-url": "3.1.0",
     "glob": "^11.0.0",
     "husky": "^8.0.0",
     "jest": "^29.7.0",
