@@ -31,13 +31,17 @@ export interface ResourceOptions {
     useCORS: boolean;
     allowTaint: boolean;
     proxy?: string;
+    customIsSameOrigin?: (this: void, src: string, oldFn: (src: string) => boolean) => boolean | Promise<boolean>;
 }
 
 export class Cache {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     private readonly _cache: { [key: string]: Promise<any> } = {};
 
-    constructor(private readonly context: Context, private readonly _options: ResourceOptions) {}
+    constructor(
+        private readonly context: Context,
+        private readonly _options: ResourceOptions
+    ) {}
 
     addImage(src: string): Promise<void> {
         const result = Promise.resolve();
@@ -61,7 +65,10 @@ export class Cache {
     }
 
     private async loadImage(key: string) {
-        const isSameOrigin = CacheStorage.isSameOrigin(key);
+        const isSameOrigin: boolean =
+            typeof this._options.customIsSameOrigin === 'function'
+                ? await this._options.customIsSameOrigin(key, CacheStorage.isSameOrigin)
+                : CacheStorage.isSameOrigin(key);
         const useCORS =
             !isInlineImage(key) && this._options.useCORS === true && FEATURES.SUPPORT_CORS_IMAGES && !isSameOrigin;
         const useProxy =
