@@ -14,7 +14,10 @@ export class Context {
 
     private static instanceCount = 1;
 
-    constructor(options: ContextOptions, public windowBounds: Bounds) {
+    constructor(
+        options: ContextOptions,
+        public windowBounds: Bounds
+    ) {
         this.logger = new Logger({ id: this.instanceName, enabled: options.logging });
         this.cache = options.cache ?? new Cache(this, options);
     }
