@@ -2,6 +2,7 @@ import { IPropertyValueDescriptor, PropertyDescriptorParsingType } from '../IPro
 import { CSSValue } from '../syntax/parser';
 import { NumberValueToken, TokenType } from '../syntax/tokenizer';
 import { Context } from '../../core/context';
+import { angle } from '../types/angle';
 export type Matrix = [number, number, number, number, number, number];
 export type Transform = Matrix | null;
 
@@ -20,21 +21,21 @@ export const transform: IPropertyValueDescriptor<Transform> = {
             if (typeof transformFunction === 'undefined') {
                 throw new Error(`Attempting to parse an unsupported transform function "${token.name}"`);
             }
-            return transformFunction(token.values);
+            return transformFunction(_context, token.values);
         }
 
         return null;
     }
 };
 
-const matrix = (args: CSSValue[]): Transform => {
+const matrix = (_context: Context, args: CSSValue[]): Transform => {
     const values = args.filter((arg) => arg.type === TokenType.NUMBER_TOKEN).map((arg: NumberValueToken) => arg.number);
 
     return values.length === 6 ? (values as Matrix) : null;
 };
 
 // doesn't support 3D transforms at the moment
-const matrix3d = (args: CSSValue[]): Transform => {
+const matrix3d = (_context: Context, args: CSSValue[]): Transform => {
     const values = args.filter((arg) => arg.type === TokenType.NUMBER_TOKEN).map((arg: NumberValueToken) => arg.number);
 
     const [a1, b1, {}, {}, a2, b2, {}, {}, {}, {}, {}, {}, a4, b4, {}, {}] = values;
@@ -42,9 +43,32 @@ const matrix3d = (args: CSSValue[]): Transform => {
     return values.length === 16 ? [a1, b1, a2, b2, a4, b4] : null;
 };
 
+const rotate = (context: Context, args: CSSValue[]): Transform => {
+    if (args.length !== 1) {
+        return null;
+    }
+
+    const arg = args[0];
+    let radians = 0;
+
+    if (arg.type === TokenType.NUMBER_TOKEN && arg.number === 0) {
+        radians = 0;
+    } else if (arg.type === TokenType.DIMENSION_TOKEN) {
+        radians = angle.parse(context, arg);
+    } else {
+        return null;
+    }
+
+    const cos = Math.cos(radians);
+    const sin = Math.sin(radians);
+
+    return [cos, sin, -sin, cos, 0, 0];
+};
+
 const SUPPORTED_TRANSFORM_FUNCTIONS: {
-    [key: string]: (args: CSSValue[]) => Transform;
+    [key: string]: (context: Context, args: CSSValue[]) => Transform;
 } = {
     matrix: matrix,
-    matrix3d: matrix3d
+    matrix3d: matrix3d,
+    rotate: rotate
 };
