@@ -110,6 +110,16 @@ const rgb = (_context: Context, args: CSSValue[]): number => {
         return pack(r, g, b, a);
     }
 
+    // Handle modern CSS syntax with / separator: rgb(r g b / alpha)
+    // tokens[0] = r, tokens[1] = g, tokens[2] = b, tokens[3] = '/', tokens[4] = alpha
+    if (tokens.length === 5 && tokens[3].type === TokenType.DELIM_TOKEN && tokens[3].value === '/') {
+        const r = getTokenColorValue(tokens[0], 0);
+        const g = getTokenColorValue(tokens[1], 1);
+        const b = getTokenColorValue(tokens[2], 2);
+        const a = getTokenColorValue(tokens[4], 3);
+        return pack(r, g, b, a);
+    }
+
     return 0;
 };
 
@@ -165,10 +175,10 @@ const _color = (context: Context, args: CSSValue[]) => {
                         part.type === TokenType.FUNCTION && part.name === 'calc'
                             ? parseCalc(part.values)
                             : isNumberToken(part)
-                            ? part.number
-                            : part.type === TokenType.DELIM_TOKEN || isIdentToken(part)
-                            ? part.value
-                            : '';
+                              ? part.number
+                              : part.type === TokenType.DELIM_TOKEN || isIdentToken(part)
+                                ? part.value
+                                : '';
                 }
                 expression += ')';
                 return expression;
@@ -193,8 +203,8 @@ const _color = (context: Context, args: CSSValue[]) => {
                 tokens[1].type === TokenType.FUNCTION
                     ? tokens[1].name
                     : isIdentToken(tokens[1]) || tokens[1].type === TokenType.HASH_TOKEN
-                    ? 'rgb'
-                    : 'unknown',
+                      ? 'rgb'
+                      : 'unknown',
             to_colorspace = isIdentToken(tokens[2]) ? tokens[2].value : 'unknown';
 
         let from =
